functions/requirements.txt
firebase-functions>=0.1.0,<1.0.0
firebase-admin>=6.5.0,<7.0.0
google-generativeai>=0.8.0,<1.0.0
google-cloud-firestore>=2.16.0,<3.0.0
google-cloud-logging>=3.10.0,<4.0.0
pydantic>=2.9.2,<3.0.0
